refactor : 안쓰는 컴포넌트 삭제, 컴포넌트별 역할 기재, 적절한 폴더로 컴포넌트이동
