+++ app/(root)/(menubar-exist)/post/[postId]/components/BotContainer.tsx
@@ -9,6 +9,7 @@ interface BotContainerProps {
   onCancelReply: () => void;
 };
 
+// 디테일 게시물 페이지의 하단 영역(채팅바)
 export default function BotContainer({ onAddComment, onAddReply, replyUser, onCancelReply }: BotContainerProps) {
   return (
     <div className="sticky bottom-0">
