+++ hooks/useGoToProfile.ts
@@ -1,6 +1,7 @@
 import { useRouter } from 'next/navigation';
 import useLoggedInUserData from './user/useLoggedInUserData';
 
+// 프로필 클릭 시 마이페이지 or 프로필페이지로 이동
 export const useGoToProfile = () => {
   const router = useRouter();
   const { userData } = useLoggedInUserData();
